// src/app/types/Car.ts
export interface Car {
  id: number;
  brand: string;
  model: string;
  year: number;
  price: number;
  marketPrice: number;
  mileage: number;
  engineVolume: number;
  gearbox: string;
  fuel: string;
  region: string;
  city: string;
  daysInSale: number;
  sellerCarCount: number;
  description: string;
  url: string;
  image: string;
  lastUpdate: string;
}